@@ -1,6 +1,6 @@
 {
   "name": "alarm_clock",
-  "version": "0.1.0",
+  "version": "0.1.5",
   "description": "Alarm Clock App",
   "main": "main.es6",
   "scripts": {
@@ -9,8 +9,12 @@
     "local": "./node_modules/webpack/bin/webpack.js && node --max-old-space-size=512 main.es6 -lv",
     "alarm": "./node_modules/webpack/bin/webpack.js && node --max-old-space-size=512 main.es6 -va",
     "watch": "./node_modules/webpack/bin/webpack.js --watch",
-    "test": "./node_modules/eslint/bin/eslint.js . --ext .jsx,.es6"
+    "test": "./node_modules/eslint/bin/eslint.js . --ext .jsx,.es6",
+    "precommit": "echo Hello_World"
   },
+  "pre-commit": [
+    "precommit"
+  ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/KyleKing/My-Programming-Sketchbook.git"
@@ -22,50 +26,49 @@
   },
   "homepage": "https://github.com/KyleKing/My-Programming-Sketchbook/tree/master/Raspberry_Pi",
   "dependencies": {
-    "babel-register": "^6.26.0",
-    "body-parser": "^1.18.3",
-    "child_process": "^1.0.2",
-    "cli-color": "^1.3.0",
-    "commander": "^2.18.0",
-    "connect-flash": "^0.1.1",
-    "cookie-parser": "^1.4.3",
-    "cookie-session": "^2.0.0-beta.3",
-    "cron": "^1.4.1",
-    "debug": "^2.6.9",
-    "express": "^4.16.3",
-    "express-session": "^1.15.6",
-    "express-users": "^0.3.2",
-    "eyespect": "^0.1.10",
-    "fs-extra": "^2.1.2",
-    "interface-addresses": "^0.0.5",
-    "karma": "^2.0.0",
-    "method-override": "^2.3.10",
-    "moment": "^2.22.2",
-    "nedb": "^1.8.0",
-    "node-pid-controller": "^1.0.1",
-    "onoff": "^1.2.0",
-    "passport": "^0.3.2",
-    "pi-blaster.js": "^0.1.3",
-    "python-shell": "^0.4.0",
-    "react": "^15.6.2",
-    "react-dom": "^15.6.2",
-    "schedule": "^0.1.0",
-    "serialport": "^4.0.1",
-    "socket.io": "^1.7.4"
-  },
-  "devDependencies": {
     "babel-core": "^6.26.3",
     "babel-eslint": "^7.2.3",
     "babel-loader": "^6.4.1",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^14.1.0",
+    "babel-register": "^6.26.0",
+    "body-parser": "^1.18.3",
+    "child_process": "^1.0.2",
+    "cli-color": "^1.4.0",
+    "commander": "^2.18.0",
+    "connect-flash": "^0.1.1",
+    "cookie-parser": "^1.4.3",
+    "cookie-session": "^2.0.0-beta.3",
+    "cron": "^1.5.0",
+    "debug": "^4.1.0",
+    "express": "^4.16.4",
+    "express-session": "^1.15.6",
+    "express-users": "file:../Pull Requests/express-users",
+    "eyespect": "^0.1.10",
+    "fs-extra": "^7.0.0",
+    "interface-addresses": "^0.0.5",
+    "karma": "^3.1.1",
+    "method-override": "^3.0.0",
+    "moment": "^2.22.2",
+    "nedb": "^1.8.0",
+    "node-pid-controller": "^1.0.1",
+    "onoff": "^3.2.2",
+    "passport": "^0.4.0",
+    "pi-blaster.js": "^0.1.3",
+    "prop-types": "^15.6.2",
+    "python-shell": "^1.0.6",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
+    "schedule": "^0.5.0",
+    "serialport": "^7.0.2",
+    "socket.io": "^2.1.1",
+    "webpack": "^2.7.0"
+  },
+  "devDependencies": {
+    "eslint": "^5.8.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.3",
-    "pre-commit": "^1.1.3",
-    "webpack": "^2.7.0",
-    "webpack-dev-server": "^2.11.3"
+    "eslint-plugin-react": "^7.11.1"
   }
 }
